Use modules
MutualInduction, Joint, and MkAll are now modules, so they can be
imported in other modules.
MutualInduction also now uses Vectors in place of Arrays to retain the
fact that the number of targeted subgoals is known after parsing.
This also means that a new error will be thrown if there are more
targets than goals, rather than the extra targets being ignored.